+++ src/style/font.css
@@ -53,6 +53,46 @@
   font-style: normal;
 }
 
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-ExtraBold.ttf') format('ttf');
+  font-weight: 800;
+}
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-Bold.ttf') format('ttf');
+  font-weight: 700;
+}
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-SemiBold.ttf') format('ttf');
+  font-weight: 600;
+}
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-Medium.ttf') format('ttf');
+  font-weight: 500;
+}
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-Regular.ttf') format('ttf');
+  font-weight: 400;
+}
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-Thin.ttf') format('ttf');
+  font-weight: 300;
+}
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-Light.ttf') format('ttf');
+  font-weight: 200;
+}
+@font-face {
+  font-family: 'Poppin';
+  src: url('../assets//fonts/Poppins-ExtraLight.ttf') format('ttf');
+  font-weight: 100;
+}
 /* @font-face {
   font-family: 'Noto Sans KR';
   src: url(assets/fonts/NotoSansKR-Black.ttf) format('truetype');
